test: extract showTransitiveDependencies coverage to dedicated class
Keep ModuleGraphPluginFunctionalTest under detekt LargeClass after
the issue #69 functional coverage was added.
