+++ plugin-build/modulegraph/src/test/java/dev/iurysouza/modulegraph/ModuleGraphPluginFunctionalTest.kt
@@ -652,99 +652,6 @@ class ModuleGraphPluginFunctionalTest {
         }
     }
 
-    @Test
-    fun `showTransitiveDependencies false shows only direct dependencies of root modules`() {
-        settingsFile.writeText(
-            """
-                plugins { id("$MODULEGRAPH_PACKAGE.settings") }
-                rootProject.name = "test"
-                include(":app")
-                include(":feature")
-                include(":core")
-            """.trimIndent(),
-        )
-        File(testProjectDir, "build.gradle.kts").writeText(
-            """
-                moduleGraphConfig {
-                    heading.set("### Dependency Diagram")
-                    readmePath.set("${readmeFilePath()}")
-                    rootModulesRegex.set(".*:app$")
-                    showTransitiveDependencies.set(false)
-                    showFullPath.set(true)
-                }
-            """.trimIndent(),
-        )
-        writeJavaModule("app", listOf(":feature"))
-        writeJavaModule("feature", listOf(":core"))
-        writeJavaModule("core", emptyList())
-        readmeFile.writeText("### Dependency Diagram")
-
-        GradleRunner.create()
-            .withProjectDir(testProjectDir)
-            .withArguments("createModuleGraph")
-            .withPluginClasspath()
-            .build()
-
-        val content = readmeFile.readText()
-        assertTrue(content.contains(":app --> :feature"), content)
-        assertFalse(content.contains(":feature --> :core"), content)
-    }
-
-    @Test
-    fun `showTransitiveDependencies true includes transitive dependency edges`() {
-        settingsFile.writeText(
-            """
-                plugins { id("$MODULEGRAPH_PACKAGE.settings") }
-                rootProject.name = "test"
-                include(":app")
-                include(":feature")
-                include(":core")
-            """.trimIndent(),
-        )
-        File(testProjectDir, "build.gradle.kts").writeText(
-            """
-                moduleGraphConfig {
-                    heading.set("### Dependency Diagram")
-                    readmePath.set("${readmeFilePath()}")
-                    rootModulesRegex.set(".*:app$")
-                    showTransitiveDependencies.set(true)
-                    showFullPath.set(true)
-                }
-            """.trimIndent(),
-        )
-        writeJavaModule("app", listOf(":feature"))
-        writeJavaModule("feature", listOf(":core"))
-        writeJavaModule("core", emptyList())
-        readmeFile.writeText("### Dependency Diagram")
-
-        GradleRunner.create()
-            .withProjectDir(testProjectDir)
-            .withArguments("createModuleGraph")
-            .withPluginClasspath()
-            .build()
-
-        val content = readmeFile.readText()
-        assertTrue(content.contains(":app --> :feature"), content)
-        assertTrue(content.contains(":feature --> :core"), content)
-    }
-
-    private fun writeJavaModule(path: String, projectDeps: List<String>) {
-        val deps = projectDeps.joinToString("\n") {
-            "    implementation(project(\"$it\"))"
-        }
-        File(testProjectDir, "$path/build.gradle.kts").apply {
-            parentFile.mkdirs()
-            writeText(
-                """
-                    plugins { java }
-                    dependencies {
-                $deps
-                    }
-                """.trimIndent(),
-            )
-        }
-    }
-
     /**
      * This is for Windows compatibility, as the path is used in the build file
      */
